ci: add prose linting with vale
This is useful to smoothen the review of submissions and the overall
quality of the content.

Vale is able to use certain comprehensive style guides (e.g. Redhat,
Microsoft, Google). This PR just uses merely the base check, which
already produces some good findings.

We can decide on a more opionated style guide later.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>