demos: fix SSH and operator demo documentation
- Fix broken links
- Update links to old branches to new branches
- Add worker node label command
- FIx typos
- Add `--progress=plain` to `docker build` so the user can see the key
  fingerprint
- Remove the "adaptation" warning because they have been tested and
verified now

I haven't tried the kcli setup example nor the crictl configuration for
port forwarding.

The `katadocker/ccv0-ssh` demo image didn't work for me as explained in
the instructions because they don't explain how to set up an
attestation-agent. We can address that part in a follow-up PR.

Signed-off-by: Iago López Galeiras <iagol@microsoft.com>