docs: update trust model
Remove incomplete threat vectors page.
We can add this back once there is some content,
but I don't quite see how it fits.

Rename personas to cloud-native-personas.

Update main trust model document to include a plain description
of the trust model rather a more generic definition of a trust model
and goals for the trust model.

The new text is mainly an overview. We may want to add something
more empirical in the future, but this should help users understand
what's going on.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>