features: add page for attestation api
Describe how to enable and use the ASR to get attestation evidence.
Also add a section about the evidence factory considerations.
That is too complex a topic to get to the bottom of here,
and we don't actually have a great solution to it until init-data
support lands, but hopefully the note is useful.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>