docs: fix old information in getting-started guide
The simple workload instructions were copied from the quickstart guide
in GitHub. There are a few issues.

They refer to an old method of checking that the container is not
present on the host. Remove this. There is another method for checking,
but it's probably too advanced to live on this page.

They link to the guides page, which doesn't exist on the website.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>