attestation: remove old documentation
Now that we have new user-facing docs, let's get rid of the older docs
that were copied from Github.

Some of this content may return in the future when we have more advanced
informatio on the website, but for now we need to focus on users.

I would like to have a listing of all the configuration options on the
website at some point, but for the time being let's simplify things.
These docs are already out-of-date anyway.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>