docs: add built-in policies to policy docs
The KBS Client now has a couple built-in policies. This allows users to
specify simply policies without having to create a file. Add this
information to the docs.

Also, switch from checking that the status isn't contraindicated to
checking that it is affirming. This is a little bit safer and probably
what we want our users to adopt.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>