docs: fix broken links
There are a couple broken links in one of our older blog posts. The link
to the RVPS predates our refactor of the attestation docs.

The link to the CAA doesn't have a direct replacement, so link to the
Azure example with CAA, since that is what is being discussed.

Signed-off-by: Tobin Feldman-Fitzthum <tfeldmanfitz@nvidia.com>